-/- I've (finally!) moved DrawPixmap() (ex-DrawSprite()) from iGraphics2D to     iGraphics3D interface. From now on iGraphics2D interface is absolutely     independent of the 3D driver and can be used separately. All the mess     regarded to getting a pointer to bitmap data has vanished. Now DrawPixmap()     extracts data from texture handle exactly same way how DrawPolygon,     DrawPolygonFX and all others do.
    I beg pardon from Direct3D and Glide maintainers since I broke Direct3D
    sprites again. I promise to never break it again, at least in the nearest
    year :-)

    I wrote a related comment in respective files near the end (g3dglide.cpp
    and d3d_g3d.cpp) - look there for more information.

    Software driver works and OpenGL driver should work too (at least it
    compiles okay).

-/- Fixed a minor bug with texture transparency in MazeD.


git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@3243 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csws/cswstex.h
@@ -40,7 +40,7 @@ class csWSTexture
   // Will be this texture used for 3D and/or for 2D operations?
   int Flags;
   // Red,Green and Blue components of transparent color
-  int tr, tg, tb;
+  UByte tr, tg, tb;
   // has texture transparent areas?
   bool IsTransp;
   // Has transparent color changed?
